# docs/Swagger/swagger.json
{"openapi":"3.0.1","info":{"title":"springdoc-openapi","description":"springdoc-openapi swagger-ui","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Monitoring","description":"데이터 페브릭의 모니터링을 위한 API로써, 서비스들의 저장소 유형 요약, 서비스들의 연결 상태 요약, 서비스들의 평균 응답 시간, 데이터 모델 추천 순위, 데이터 모델 등록 현황, 히스토리를 제공하는 도구입니다."}],"paths":{"/api/v1/monitoring/statusCheck":{"get":{"tags":["Monitoring"],"summary":"Status Check","description":"모니터링의 상태를 확인하기 위한 API입니다.","operationId":"statusCheck","responses":{"200":{"description":"OK"}}}},"/api/v1/monitoring/runSchedule":{"get":{"tags":["Monitoring"],"summary":"Manual Schedule","description":"Monitoring 정보 수집을 위해 자동적으로 돌아가는 Schduler를 수동으로 돌리기 위한 API","operationId":"Schedule","parameters":[{"name":"userName","in":"query","description":"스케쥴링을 시도한 사용자 이름에 대한 매개변수","required":true,"schema":{"type":"string","example":"admin"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/monitoring/responseTime":{"get":{"tags":["Monitoring"],"summary":"Response Time","description":"모든 서비스들의 평균 응답 시간을 얻기 위한 API","operationId":"responseTime","parameters":[{"name":"orderByAsc","in":"query","description":"평균 응답 시간의 내림차순 혹은 오름차순을 정하기 위한 매개변수","required":false,"schema":{"type":"boolean","example":true}},{"name": "pageNumber","in":"query","description":"요청된 데이터의 페이지 번호를 위한 매개변수","required":false,"schema":{"minimum":0,"type":"int","example":0}},{"name": "pageSize","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"minimum":1,"type":"int","example":5}}],"responses":{"200":{"description":"모든 서비스들의 평균 응답 시간 정보","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServicesConnect"}}}}}}}},"/api/v1/monitoring/responseTime/{serviceID}":{"get":{"tags":["Monitoring"],"summary":"Target Response Time","description":"특정 서비스의 응답 시간 히스토리를 얻기 위한 API","operationId":"targetResponseTime","parameters":[{"name":"serviceID","in":"path","description":"응답 시간 히스토리를 얻을 특정 서비스의 아이디","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"요청된 데이터의 페이지 번호를 위한 매개변수","required":false,"schema":{"type":"int","example":0}},{"name":"size","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"type":"int","example":5}}],"responses":{"200":{"description":"특정 서비스들의 응답 시간 히스토리 정보","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServicesConnect"}}}}}}}},"/api/v1/monitoring/models":{"get":{"tags":["Monitoring"],"summary":"Ranking of Voted Model","description":"데이터 모델의 추천 순위를 위한 API","operationId":"model","parameters":[{"name": "pageSize","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"minimum":1,"type":"int","example":5}}],"responses":{"200":{"description":"데이터 모델의 추천 순위 정보","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Services"}}}}}},"deprecated":true}},"/api/v1/monitoring/eventHistory":{"get":{"tags":["Monitoring"],"summary":"Event History","description":"히스토리를 위한 API","operationId":"eventHistory","parameters":[{"name":"size","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"minimum":1,"type":"int","example":5}}],"responses":{"200":{"description":"히스토리 정보","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Services"}}}}}}}},"/api/v1/monitoring/eventHistory/{serviceID}":{"get":{"tags":["Monitoring"],"summary":"Tagret Event History","description":"특정 서비스의 히스토리를 위한 API","operationId":"targetEventHistory","parameters":[{"name":"serviceID","in":"path","description":"히스토리를 얻을 특정 서비스의 아이디","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"요청된 데이터의 페이지 번호를 위한 매개변수","required":false,"schema":{"minimum":0,"type":"int","example":0}},{"name":"size","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"minimum":1,"type":"int","example":5}}],"responses":{"200":{"description":"특정 서비스의 히스토리 정보","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}}}}}},"/api/v1/monitoring/connectStatus":{"get":{"tags":["Monitoring"],"summary":"Connect Status","description":"모든 서비스들의 연결 상태를 위한 API","operationId":"connectStatus","responses":{"200":{"description":"서비스들의 연결 상태에 대한 정보","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectStatusResponse"}}}}}}},"/api/v1/monitoring/connectStatus/{serviceID}":{"get":{"tags":["Monitoring"],"summary":"Target Connect Status","description":"특정 서비스의 연결 상태를 위한 API","operationId":"targetConnectStatus","parameters":[{"name":"serviceID","in":"path","description":"연결 상태 히스토리를 얻을 서비스의 아이디","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"요청된 데이터의 페이지 번호를 위한 매개변수","required":false,"schema":{"minimum":0,"type":"int","example":0}},{"name":"size","in":"query","description":"한 페이지에 표시할 데이터의 수를 나타내는 매개변수","required":false,"schema":{"minimum":1,"type":"int","example":5}}],"responses":{"200":{"description":"특정 서비스의 연결 상태에 대한 히스토리 정보","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Services"}}}}}}},"/api/v1/monitoring/Test":{"get":{"tags":["Monitoring"],"operationId":"test","responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ServicesConnect":{"type":"object","properties":{"executeAt":{"type":"string","description":"응답 시간 측정을 위한 실행 시작 시간","format":"date-time"},"executeBy":{"type":"string","description":"응답 시간 측정을 요청한 사용자의 이름","example":"admin"},"queryExecutionTime":{"type":"integer","description":"평균 응답 시간","format":"int64"},"serviceName":{"type":"string","description":"응답 시간을 측정한 서비스의 이름"},"serviceID":{"type":"string","description":"응답 시간을 측정한 서비스의 UUID","format":"uuid"}}},"Services":{"type":"object","properties":{"serviceID":{"type":"string","description":"Service의 UUID","format":"uuid"},"name":{"type":"string","description":"Service의 이름"},"serviceType":{"type":"string","description":"Service의 타입","example":"Mysql"},"ownerName":{"type":"string","description":"소유자의 이름","example":"admin"},"createdAt":{"type":"string","description":"Service가 생성된 날짜","format":"date-time"},"deleted":{"type":"boolean","description":"Service의 Hard Delete 유무"},"connectionStatus":{"type":"boolean","description":"Service의 Connection 가능 여부"},"connects":{"type":"array","items":{"$ref":"#/components/schemas/ServicesConnect"}},"histories":{"type":"array","items":{"$ref":"#/components/schemas/ServicesHistory"}}}},"ServicesHistory":{"type":"object","properties":{"updateAt":{"type":"string","description":"이벤트가 발생한 시간","format":"date-time"},"event":{"type":"string","description":"이벤트 명"},"serviceID":{"type":"string","description":"이벤트가 발생한 서비스의 UUID","format":"uuid"}}},"ConnectStatusResponse":{"type":"object","properties":{"total":{"type":"integer","format":"int64"},"connected":{"type":"integer","format":"int64"},"disConnected":{"type":"integer","format":"int64"}}}}}}